date, time issues - temporarily fixed!
