+++ functions/Astronomicaltime/astronomicaltime.js
@@ -3,13 +3,9 @@ import {getSunrise, getSunset} from 'sunrise-sunset-js';
 
 export const getSuntimings = setTime => {
   Geolocation.getCurrentPosition(info => {
-    setTime(prev => ({
-      ...prev,
-      sunrise: getSunrise(info.coords.latitude, info.coords.longitude),
-    }));
-    setTime(prev => ({
-      ...prev,
+    setTime(() => ({
       sunset: getSunset(info.coords.latitude, info.coords.longitude),
+      sunrise: getSunrise(info.coords.latitude, info.coords.longitude),
     }));
   });
 };
