+++ screens/Register.js
@@ -39,7 +39,7 @@ const Register = ({navigation}) => {
     // handle wrong input
     if (!validation(input)) {
       //! todo: uncomment!
-      // return;
+      return;
     } else {
       setErr('');
     }
